+++ src/components/Categories.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Categories() {
+  return (
+    <>
+      <h2>Categories</h2>
+      <button type="button" className="categorie">check status</button>
+    </>
+  );
+}
+
+export default Categories;
